+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.1] - 2017.02.08
+
+- Use the capture group for rgb(a) the decoration and not the complete match
+- Generate decorations only once, after the end of the text analysis
+
 ## [0.3.0] - 2017.02.04
 
 - Use Promise for background generation/update for easy chaining